describe what this is about
Why "gammarus": In Czech, this is blešivec, a small animal which can act
as a bioindicator of certain water conditions. IMHO, this is good enough
to qualify as a name for a dashboard. Besides, "gamma" and "lambda" are
both Greek letters, and "lambda" has an obvious relation with DWDM.

Change-Id: I8dc3fa5ed4ed5a0378ea424981b98cb27938781c