OSA: do not zoom-in too much into the noise
Instead of showing a detailed view of the OCM noise when there's total
darkness, use these defaults to make the random noise appear reasonably
flat. This has been brought up several times during the demos.

When the actual values are outside of the range, the chart will still
adapt so that nothing gets cut off.

Change-Id: If0a5306acda09b411dc223f3289252fec8b03689