Update patch set 1
Patch Set 1:

> changing "fill: " to true makes them the right color again. It's probably connected to this: https://github.com/chartjs/Chart.js/pull/6719

If I do that at the dataset level, I'm getting a filled graph (and a correct graph color). This is apparently a behavior change introduced in 2.9.0.

> It is possible to make the label color red/blue by changing the line color and with "borderWidth: 0.2"

That works, but then the colors are different than they used to be, to the point that the orange-on-white graph for the Add/Drop is not that well readable anymore. Previously, the gray lines used to improve contrast (and they also darkened the orange color of the main line when zoomed out as a side effect).

I'll remove the JS update from this commit and make it an extra change, the rest of these changes IMHO make sense.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>