Update (and unify) Graphs.JS
This update changes rendering of graph legend. Previously, our code
relied on "graph lines" and "graph points" of different colors (the
"points" were colored, while the connecting line was implicitly dark).
An upstream change in Graphs.JS changed this, they now use the line
color for legend fill area, which means that we will have a gray legend.

There are many possible ways of fixing this. After experimenting with
this, I decided not to go the legend.labels.usePointStyle way, and
instead I'm just tweaking several colors in an ad-hoc style. Well, it
works.

Change-Id: I720577800b7108a091d469f2e6801da522073305