Update Graphs.JS
This update changes rendering of graph legend. Previously, our code
relied on "graph lines" and "graph points" of different colors (the
"points" were colored, while the connecting line was implicitly dark).
An upstream change in Graphs.JS changed this, they now use the line
color for legend fill area, which means that we will have a gray legend
unless we make extra changes. And gray legend for two graphs of
different colors is not really awesome.

There are many possible ways of fixing this. After experimenting with
this, I decided not to go the legend.labels.usePointStyle way, and
instead I'm just tweaking several colors, line stroke widths and point
diameters in a rather ad-hoc style. Well, it works; the key was really
amking the orange a little bit darker and also ensuring that the orange
stroke width is a wee bit wider.

Change-Id: I720577800b7108a091d469f2e6801da522073305
Bug: https://github.com/chartjs/Chart.js/issues/6644
Bug: https://github.com/chartjs/Chart.js/pull/5621
Bug: https://github.com/chartjs/Chart.js/issues/5375